do not short circuit waitForTransform timeout when running inside pytf. Fixes #102
roscpp is not initialized inside pytf which means that ros::ok is not
valid. This was causing the timer to abort immediately.

This breaks support for pytf with respect to early breaking out of a loop re #26.
This is conceptually broken in pytf, and is fixed in tf2_ros python implementation.
If you want this behavior I recommend switching to the tf2 python bindings.