Fixing do_transform_cloud for python
The previous code was not used at all (it was a mistake in the __init__.py so
the do_transform_cloud was not available to the python users).

The python code need some little correction (e.g there is no method named
read_cloud but it's read_points for instance, and as we are in python we can't
use the same trick as in c++ when we got an immutable)

I added a simple test for the point_cloud python version too.