Switch tf2_eigen to use package.xml format 2. (#216)
* Switch tf2_eigen to use package.xml format 2.

This is to modernize it a bit to make it more compatible
with ROS 2.  While we are in here, remove the run/exec dependency
on cmake_modules, which doesn't seem to be needed at runtime.

Signed-off-by: Chris Lalancette <clalancette@osrfoundation.org>

* Make geometry_msgs and tf2 <depend>

This makes sure they are exec_depend, build_depend, and
build_export_depend, all three of which are needed.  Also
make eigen a build_export_depend, since it is included in
the exported header file.

Signed-off-by: Chris Lalancette <clalancette@osrfoundation.org>